api/requirements.txt
fastapi
uvicorn[standard]
httpx
gunicorn
python-multipart
orjson
python-dotenv
